refactor: 승패출력 오류 수정 및 파일 분류
사용자 승리와 컴퓨터 승리 동시 출력 오류 수정/ 함수 특성별 파일 분류
